Rename: regions.rs => region.rs
Closes #118.
